pyproject.toml
[project]
name = "titan-sspr"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "flask>=3.0",
    "flask-session>=0.8",
    "flask-limiter>=3.5",
    "flask-wtf>=1.2",
    "msal>=1.28",
    "requests>=2.32",
    "ldap3>=2.9",
    "sqlalchemy>=2.0",
    "argon2-cffi>=23.1",
    "gunicorn>=22.0",
    "python-dotenv>=1.0",
    "werkzeug>=3.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "pytest-flask>=1.3",
]

[tool.hatch.build.targets.wheel]
packages = ["app"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
